Add API endpoint for adding an item to a cart
Endpoint /api/cart/add takes a POST request with urlencoded data
Data should include productId=(id of the product being added)

Will add this product to the cart that exists for the user associated
with the `id` cookie present on the POST request. If no cart exists for
that cookie value yet, a new cart (i.e., an entry in the `carts` db
table) will be created.

Code is still at 100% coverage.