Update address validation endpoint, re-integrate with frontend
Address validation is done by preventing form submission, grabbing
all the address-related form variables with jQuery, sending them
to the backend validation endpoint via GET request, then handling the
result of the validation in the GET callback handler.

We'll probably want the callback handler to include, at minimum,
something that displays an indication that the address is valid
(or not) and something that allows the entire checkout form to
submit once the address has been validated.